Split SAX handler for ImageIndex.*.xml into a separate class
